Replaced the custom PydidasDataViewerFrame with a simplified silx DataViewerFrame which only replaced the 2d view.
