+++ pydidas/gui/frames/builders/data_browsing_frame_builder.py
@@ -31,6 +31,7 @@ __all__ = ["DataBrowsingFrameBuilder"]
 import qtawesome as qta
 from qtpy import QtWidgets
 
+# from ....widgets.silx_plot import PydidasDataViewerFrame
 from ....core.constants import POLICY_EXP_EXP
 from ....widgets.factory import SquareButton
 from ....widgets.framework import BaseFrame
@@ -39,7 +40,7 @@ from ....widgets.selection import (
     Hdf5DatasetSelector,
     RawMetadataSelector,
 )
-from ....widgets.silx_plot import PydidasDataViewerFrame
+from ....widgets.silx_plot.silx_data_viewer import SilxDataViewer
 
 
 class DataBrowsingFrameBuilder:
@@ -108,7 +109,7 @@ class DataBrowsingFrameBuilder:
         )
         frame.add_any_widget(
             "viewer",
-            PydidasDataViewerFrame(),
+            SilxDataViewer(),
             parent_widget="viewer_and_filename",
             gridPos=(1, 0, 1, 2),
         )
